# Hand-code remote-write protobuf