… storage/remote/codec_test.go
… storage/remote/marshal_test.go
… storage/remote/write_handler_test.go