Draft, as you may not want to add the complexity, but I thought it was worth seeing the performance achievable.

```
name          old time/op    new time/op    delta
SampleSend-4    1.70ms ± 0%    1.12ms ± 4%  -34.44%  (p=0.016 n=4+5)

name          old alloc/op   new alloc/op   delta
SampleSend-4    47.4kB ± 0%    29.6kB ± 1%  -37.57%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
SampleSend-4       140 ± 0%       110 ± 0%  -21.43%  (p=0.008 n=5+5)
```

The code is pretty much as generated by the proto compiler, but working off the data structure created in #9934. We benefit from not creating the intermediate structs that protoc normally uses. 
As an additional step I re-wrote some of the varint code to go directly to the number of 7-bit words required, and optimise for small numbers

Another step I experimented with was merging external labels at the point of marshalling rather than creating an additional slice for each series in `processExternalLabels` called from `StoreSeries`. This does reduce heap size but increases CPU about 15% on every send, so I left it out.