@bboreham 2021-12-23
@csmarchbanks do you have a view on this?