@cstyan 2022-01-06
I personally wouldn't want to maintain the marshalling code, so it's up to Chris (or if Bryan is offering to keep with maintenance of the marshalling code). Given that this version of the protocol is pretty stable and likely won't change until a new protocol version I think we're fine to move forward, the improvements shown in the benchmark are pretty compelling.