@bboreham 2022-01-07
>  if Bryan is offering to keep with maintenance of the marshalling code

Sure, I can do that.  (Assuming there is some value in a random person on the internet committing to do something)